Use requirement checks and not anvil version
Instead of using the anvil version in the bootstrap
identifer file use the checksum of the pip and pkg
requirement files instead which is more accurate then
the anvil version. Since pkg/pip dependencies for anvil
sometimes change inside a version this checksum approach
is more resilent to those types of changes as well as
being resilent between versions when the dependency list
doesn't change.