Yaml interpolation cleanups + addons
Move the if statements for the 'auto' key
into a dictionary of keys => functors which
will be looked up whenever a yaml 'auto' special
value is encountered. This allows others to extend
this if desired without having to add new if
statements on.

Cleanup some type conversions and variable naming also.