pylint should only fail test on fatal error
pylint will exit non-zero on any type of issue it uncovers, such
as a code convention issue.  We only want to fail the tests if
pylint discovers fatal issues.

There are currently only a few errors and dozens of warnings.  If
we can get those cleaned up then we can cause the test to fail
when new errors (or warnings) are introduced.