Robust handling for existing git repos
If a repository was downloaded previously, it may contain
local changes. Do git clean -xdfq and git reset --hard.

If a repository was downloaded previously, it may contain
a branch with the smae name we want to use. To be compatible
with old git, we drop it and create again.

Change-Id: Ib46328e9cddc8d1e58833c08b65679b1b61a891d