Move multipip and associated to tools + docs.
Use the readme.rst for multipip for other tools and
place multipip and the new tools under tools since they
are anvil tools and should be under that directory.

Adjust the code to look under that directory when finding
py2rpm and multipip.

Remove some tools which are no longer needed/used/make sense.

Change-Id: If3a0d627c42ef0374f764ad9e276424b43f29a66