Move package management to yum tool
This commit introduces new tool named YYOOM that handles package
management. It uses yum python module (and so is GPLv2+ licensed)
and print work result in JSON to stdout.

Anvil uses it to install packages and query which packages are
install or available. We also log all installed packages (including
deps) via tracewriter, which allows to use tracereader for
clean and complete uninstall.

Fixes: bug 1189707
Change-Id: Ib6d13b2dc816a3d2f8875aa23779e34fa685cd31