Rename quantum to neutron mechanically
Quantum is now called neutron, so we should do the renaming, too.
This commit is the result of simple mechanical search-and-replace
followed by renaming files, followed by reordering imports to make
pep8 happy.

Building of binary RPMs is still broken, as well as many other things,
but this will be fixed in separate changes.

Partial-bug: #1198041
Co-authored-by: Joshua Harlow <harlowja@gmail.com>
Change-Id: Idd1cce74309e67fd6bbb1348f40fcc9946703fc2