Allow for running tests via testr.
Some projects in openstack are using nose,
some are using testr so attempt to select
the command used by first checking an option
that the component may have set. If that is
not set then attempt to see if the testr
configuration files exists, otherwise select
nose as the backup.

Change-Id: I8e06235f54dded11093d747f161dc4f95b7f6cba