Show status before running py2rpm+yumfind.
This stage can actually take a while to finish
so before starting it, it is pretty useful to
tell the user that we are starting the bootstrap
of rpms so they know what is going on. This is
especially relevant if yumfind has to internally
update repos (like epel, which can take a while
to list).

Change-Id: Iaf907681a9b28e197269a5e88ef05bc6d88750c8