Anvil is removing its dependencies again.
On install it appears the python-keyring
package is installed (its a newer version
that what gets bootstrapped). Then this
package gets recorded and then will be
removed on uninstall. We don't want to
remove this so ensure that we keep all
our dependencies even if they have been
upgraded underneath us.

Fixes bug 1211647

Change-Id: I86a9cdfc66ba3f57df266141f993d152c1d93dbe