Put RPMS build on bootstrap to anvil-deps repo
RPMs build on bootstrap (before any python code gets a chance to run)
are now put to temporary but well-defined locations in ./.bootstrap
directory, and then moved to appropriate place at the beginning of
prepare action, when openstack root dir is already known.

So, said RPMs are not lost, but put in place from which they could be
re-installed or installed to different system.

Implements: blueprint bootstrap-pip-rpm
Change-Id: I41dded0ecd8d887b540aea87d0226be03508c3cb