Fix loading of python names in dependency resolver
When dependency resolver is created before package directories
are downloaded, python_names list in it is empty. This leads
to several subtle errors like missing epoch map and thus invalid
epoch in component packages build with py2rpm.

To address the issue, we try to reload python names and package
dirs until they are not empty.

Change-Id: I014a561fad4308b2d4fbef9d628a2ac175997426