Add original requirement when converting using py2rpm
Output the original requirement a rpm requirement is
being created for and use this original requirement
output to make sure all the requirements that are
being output in rpm form actually match the ones that
were requested (and blowup if this does not happen).

As py2rpm now requires six, it is added to CommonRedHat
requirements.

Co-authored-by: Ivan A. Melnikov <imelnikov@griddynamics.com>
Change-Id: I1f43b20168f3892184d849f1a8a0c982bb70e0a3