Remove usage of pbr
Anvil has conflicts with pbr when it's used to in setup.py
and its usage in anvil doesn't really seem useful or applicable
given the amount of pain it causes during packaging and later pip
usage.

So instead of using it, just switch back to using a simpler
setup.py file with usage of nosetests for testing (since we
don't need parallel tests for anvils tests) to avoid the
problems that pbr has caused.

Change-Id: I6d3c09a927434abf07fc025638a0860e44b029fa