Add a bootstrap repo file stage
Allow for fetching additional dependencies for openstack
by allowing for fetching of arbitary repo files during
the bootstrap process.

Add an initial example that allows for fetching the RDO
repo file (which provides quite a few of the same python
dependencies that anvil will build automatically if it
can not find them itself).

Change-Id: Ia51adf017d398863432ecb9d218e163235039488