Select most requested when equivalent scores
If equivalent scores, then select the one that has the
most requests for its selected version combinations over
ones that have less requests for those same version
combinations.

Also show all requirements that were chosen (instead of
just the first match of many). This makes it easy to see
why the scoring and selection was done.

Change-Id: I7abed2b7017b1f52cf7fe97284edbadd04bd8a34