Lack of nova-rpc-zmq-receiver causing build breakage
It appears that the nova-rpc-zmq-receiver is not being
included anymore in nova. This is causing the building of
the nova package to break with file not found errors.

To avoid this, check against a given commit to know whether
we should include or exclude the inclusion of the zmq-receiver
binary.

The commit that removed it is 4cd9ce0 and 1a91aac.

While at it also fix the neutron commits checks, which also
appear to have changed version scheme (pbr change again??).

Fixes bug 1300527

Change-Id: I951e77cc7af4c22e246dd5960c9ad70c239a24a1