Fix empty ini files not including DEFAULT section
Empty files and subsequent modification by adding, removing,
setting and changing keys then outputting those files seems
to cause iniparse to ignore the whole DEFAULT section. To fix
this modify the write() function to correctly output the DEFAULT
section when it exists (but does not exist in the read data).

Fixes bug 1303968

Change-Id: I1b4133bc6f876e9e5e39045f46054db13bd4634f