Automatically disable components not found in origins
Instead of blowing up with components that are existent
but not disabled, just disable them automatically if they
are present in the persona file (which most will be) but
not present in the origins file (which some may not be).

This is to accomodate persona files having more components
than an origins file (which may be ok if those components
are not actually needed).

Change-Id: Ia5c51177a58ba686cf008f2958d72543705953cb