Instead of messing with yum output just write json to a file
Instead of messing with the yum output, which makes it hard
to debug when failures occur just write the json output to
a provided filename and load this when running yyoom instead.

At the same time write the yyoom full log to a given log dir
that contains each yyoom invocations output (now with more
details!).

Change-Id: I550265bb890b4f785ddbcaa8abff03e2fe5e8983