Merge origins components
Also merge in any component origin settings so that
these options can be used through the rest of the
system.

Origin specific options are useful to be able to
target options to a specific origin/release, for
example to make a havana release have a specific
rpm option, while leaving the other origins to not
have this option.

Change-Id: I11aca326c3e715427963e0c881d78c85e17c1ffd