Remove oslo.messaging, oslo.config, pycadf as custom builds
Instead of always building these, we should just let them
be built by py2rpm and download via pip now that these packages
exist on pypi and can be fetched from there as needed.

Projects should no longer (if ever) be depending on versions which
are not published on pypi so it should be fine to do this
component adjustment.

This will also make it easier for packages with build-dependencies
on these components to actually build and install correctly since
building currently does not publish a repo after each *individual*
component is built.

Partially fixes bug 1355439

Change-Id: Id304f93b8c5ed9caaafa58a4667032a8597aef16