Ensure we pickup previously built epochs
When we have previously built a package and have its repo
accessible we should make sure that when building dependencies
that we set the correct epoch in there built package.

This change will scan the dependency which were matched via yum
and extract there epoch from there found rpm information and include
it automatically in the epoch mapping sent into py2rpm during building.
It will skip a set of epochs that should not be provided (by default
this will always skip the 0 epoch).

Note that with this change built dependencies will also inherit any
custom epochs that are found in epel or elsewhere as well...

Fixes bug 1356429

Change-Id: I0a82b41a4a734a1ecbdbd9b5fcef2b3dfb39ea04