Fix failure when only building horizon
If only the django-openstack-auth, horizon and general components are
enabled in the persona then, during the prepare phase, group 0 generates
no SRPMS. This results in no 'generated-rpms-0' file being generated.

In the build action, absence of this file was not (before this change)
equivalent to an empty version of this file - it resulted in the list of
SRPMs not being filtered down, so for group 0 the build of all SRPMs
would be attempted, which fails for dependency reasons.

This change means that absence of the generated-rpms file is equivalent
to it being present and empty.

Co-Authored-By: Ben Motz <bmotz@cray.com>
Change-Id: I1ca2d644f5fb9bca53af9da2a58c502d693cd67a