Start to add release specific specs
The number of conditionals is starting to become to many
so we can easily split these into release specific specs
and just work via these instead (and use #include as needed).

This starts that by adding subdirs with spec files that just
include the top-level specfile, and makes it so that the order
of spec file finding is to look in the release specific dir and
then look in the base folder for a spec file that matches.

Change-Id: I71b8c3bf0d6b2d7c00b487943c780dadc772a04c